win records almost working but still working with one glitch

+++ db/define_schema.rb
@@ -3,7 +3,7 @@ require_relative "./_configure"
 DB.define_table("planets")
 DB.define_column("planets", "planet", "string")
 DB.define_column("planets", "points", "integer")
-DB.define_column("planets", "win_record", "decimal")
+DB.define_column("planets", "win_record", "float")
 DB.define_column("planets", "legs_points", "integer")
 DB.define_column("planets", "smile_points", "integer")
 DB.define_column("planets", "pres_points", "integer")
